+++ CHANGELOG.md
@@ -46,6 +46,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Fixed test name for basic gets in functional tests
 - Seeders in startup will newline when there are multiple entities
 - Unicode now onlyenforced on windows (for better emoji support)
+- Fixed `isRequired` property
 
 ### Removed